Restructure File Structure
Modified file structure and CMake files.
No change in functionalities.
