+++ IReflx/src/IStarReflextor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "CommandLineParser.h"
+#include <IReflx/CommandLineParser.h>
 
 namespace ThetaStream
 {
