+++ IReflx/src/StdinReader.cpp
