+++ IReflx/src/UdpData.cpp
@@ -1,7 +1,5 @@
 #include "UdpData.h"
 
-#include <memory>
-#include <algorithm>
 #include <iterator>
